+++ packages/sdk/src/chains/Substrate/__test__/deposit.test.ts
@@ -1,33 +1,20 @@
 import { ApiPromise } from '@polkadot/api';
 import { BN } from '@polkadot/util';
-import { web3FromAddress } from '@polkadot/extension-dapp';
-import { InjectedAccountWithMeta } from '@polkadot/extension-inject/types';
 import { XcmMultiAssetIdType } from '../types';
-import { DepositCallbacksType } from '../utils/depositFns';
 
 import * as Utils from '../utils/depositFns';
 
 jest.mock('@polkadot/api');
-jest.mock('@polkadot/extension-inject/types');
-jest.mock('react');
-
-jest.mock('@polkadot/extension-dapp', () => ({
-  web3FromAddress: jest.fn().mockResolvedValue({ signer: jest.fn() }),
-}));
 
 describe('deposit', () => {
   let api: ApiPromise;
-  let currentAccount: InjectedAccountWithMeta;
   let xcmMultiAssetId: XcmMultiAssetIdType;
   let amount: string;
   let domainId: string;
   let address: string;
-  let callbacksMockFns: DepositCallbacksType;
 
   beforeEach(() => {
     api = new ApiPromise();
-
-    currentAccount = { address: 'testaddress' } as unknown as InjectedAccountWithMeta;
 
     xcmMultiAssetId = { id: 1 } as unknown as XcmMultiAssetIdType;
 
@@ -37,16 +24,16 @@ describe('deposit', () => {
 
     address = 'testaddress';
     jest.spyOn(Utils, 'calculateBigNumber').mockReturnValue(new BN(110));
-
-    callbacksMockFns = {
-      onInBlock: jest.fn(),
-      onFinalized: jest.fn(),
-      onError: jest.fn(),
-    };
   });
 
-  it('should call calculateBigNumber with correct params', async () => {
+  it('should call calculateBigNumber with correct params', () => {
     const signAndSendMockFn = jest.fn();
+    jest.spyOn(Utils, 'createMultiAssetData').mockReturnValue({
+      id: { id: 1 },
+      fun: {
+        fungible: '100000000000000000',
+      },
+    });
 
     api = {
       tx: {
@@ -58,169 +45,37 @@ describe('deposit', () => {
       },
     } as unknown as ApiPromise;
 
-    await Utils.deposit(
-      api,
-      currentAccount,
-      xcmMultiAssetId,
-      amount,
-      domainId,
-      address,
-      callbacksMockFns,
-    );
+    Utils.deposit(api, xcmMultiAssetId, amount, domainId, address);
+    expect(Utils.createMultiAssetData).toHaveBeenCalledWith(xcmMultiAssetId, api, '1000000000000');
+    jest.resetAllMocks();
+  });
+});
+
+describe('createMultiAssetData', () => {
+  let api: ApiPromise;
+  let xcmMultiAssetId: XcmMultiAssetIdType;
+  let amount: string;
+
+  beforeEach(() => {
+    api = new ApiPromise();
+
+    xcmMultiAssetId = { id: 1 } as unknown as XcmMultiAssetIdType;
+
+    amount = '1000000000000';
+
+    jest.spyOn(Utils, 'calculateBigNumber').mockReturnValue(new BN(110));
+  });
+  it('should call calculateBigNumber and return a multi-asset data object', () => {
+    api = {} as unknown as ApiPromise;
+
+    const result = Utils.createMultiAssetData(xcmMultiAssetId, api, amount);
 
     expect(Utils.calculateBigNumber).toHaveBeenCalledWith(api, amount);
-  });
-
-  it('should call web3FromAddress with correct params', async () => {
-    const signAndSendMockFn = jest.fn();
-
-    api = {
-      tx: {
-        sygmaBridge: {
-          deposit: jest.fn().mockImplementationOnce(() => {
-            return { signAndSend: signAndSendMockFn };
-          }),
-        },
+    expect(result).toEqual({
+      id: xcmMultiAssetId,
+      fun: {
+        fungible: '110',
       },
-    } as unknown as ApiPromise;
-
-    await Utils.deposit(
-      api,
-      currentAccount,
-      xcmMultiAssetId,
-      amount,
-      domainId,
-      address,
-      callbacksMockFns,
-    );
-
-    expect(web3FromAddress).toHaveBeenCalledWith(currentAccount.address);
-  });
-
-  it('should call sygmaBridge.deposit with correct params', async () => {
-    const signAndSendMockFn = jest.fn();
-
-    api = {
-      tx: {
-        sygmaBridge: {
-          deposit: jest.fn().mockImplementationOnce(() => {
-            return { signAndSend: signAndSendMockFn };
-          }),
-        },
-      },
-    } as unknown as ApiPromise;
-
-    await Utils.deposit(
-      api,
-      currentAccount,
-      xcmMultiAssetId,
-      amount,
-      domainId,
-      address,
-      callbacksMockFns,
-    );
-
-    expect(api.tx.sygmaBridge.deposit).toHaveBeenCalledWith(
-      {
-        fun: {
-          fungible: '110',
-        },
-        id: {
-          id: 1,
-        },
-      },
-
-      {
-        parents: 0,
-        interior: {
-          x2: [{ generalKey: address }, { generalKey: '0x01' }],
-        },
-      },
-    );
-  });
-
-  it('should call signAndSend', async () => {
-    const signAndSendMockFn = jest.fn();
-    api = {
-      tx: {
-        sygmaBridge: {
-          deposit: jest.fn().mockImplementationOnce(() => {
-            return { signAndSend: signAndSendMockFn };
-          }),
-        },
-      },
-    } as unknown as ApiPromise;
-
-    await Utils.deposit(
-      api,
-      currentAccount,
-      xcmMultiAssetId,
-      amount,
-      domainId,
-      address,
-      callbacksMockFns,
-    );
-
-    expect(signAndSendMockFn).toHaveBeenCalledTimes(1);
-  });
-
-  it('should call handleTxExtrinsicResult', async () => {
-    const unsub = jest.fn();
-    const signAndSendMockFn = jest
-      .fn()
-      .mockImplementationOnce((account, options, callback: (v: string) => void) => {
-        callback('someResult');
-        void Promise.resolve(unsub);
-      });
-
-    api = {
-      tx: {
-        sygmaBridge: {
-          deposit: jest.fn().mockImplementationOnce(() => {
-            return { signAndSend: signAndSendMockFn };
-          }),
-        },
-      },
-    } as unknown as ApiPromise;
-    jest.spyOn(Utils, 'handleTxExtrinsicResult').mockImplementation();
-    await Utils.deposit(
-      api,
-      currentAccount,
-      xcmMultiAssetId,
-      amount,
-      domainId,
-      address,
-      callbacksMockFns,
-    );
-    expect(Utils.handleTxExtrinsicResult).toHaveBeenCalledTimes(1);
-    expect(Utils.handleTxExtrinsicResult).toHaveBeenCalledWith(
-      api,
-      'someResult',
-      undefined,
-      callbacksMockFns,
-    );
-  });
-
-  it('should reject in case of error', async () => {
-    const signAndSendMockFn = jest.fn().mockRejectedValue(new Error('Sick error'));
-    api = {
-      tx: {
-        sygmaBridge: {
-          deposit: jest.fn().mockImplementationOnce(() => {
-            return { signAndSend: signAndSendMockFn };
-          }),
-        },
-      },
-    } as unknown as ApiPromise;
-    await Utils.deposit(
-      api,
-      currentAccount,
-      xcmMultiAssetId,
-      amount,
-      domainId,
-      address,
-      callbacksMockFns,
-    );
-    expect(callbacksMockFns.onError).toBeCalledWith(new Error('Sick error'));
+    });
   });
 });